♻️ refactor(config): extract api keys editor
Move the visual config API keys card editor into its own component file while preserving the existing blocks export.

Reduce VisualConfigEditorBlocks to the payload rule editors so the config UI can be split further in smaller pieces.

Validated with type-check, config tests, lint, and Vite production build.
